Rewrite queue processing using Reader/Writer
The Reader/Writer were added to the crate virtio-queue in version 0.12.0
This commit updates this dependency. Also it now uses a rebased fork of
the vhost repository (this was needed for it to also use the newer
virtio-queue dependency).

This commit also removes some unused imports.

Changing to using the reader revealed the following bugs:

struct virtio_gpu_get_edid was incorect for our use case - it contained
an aditional virtio_gpu_ctrl_hdr, so that meant we were trying to
read the header twice.

GetDisplayInfo also contained an additional virtio_gpu_ctrl_hdr, which
we were also trying to read twice.

The command `ResourceAttachBacking` is now parsed corectly and reported
as not implemented.

Signed-off-by: Matej Hrica <mhrica@redhat.com>