FIXME: Added a workaround in transfer_write
This works around an issue with a guest image that uses transfer_write to
updatea portion of the screen - such as a server image without GUI
displaying a console. Without this workaround most of the image would be
missing / the screen would update weirdly. GUI images that update the whole
screen area at once where not effected by this issue.

Signed-off-by: Matej Hrica <mhrica@redhat.com>